More progress including handling write correctly
